refactor: move `TeleportViewManager` in `managers` package
